requirements.txt
# GAIA/MYCEL (already available)
rag-intelligence>=0.3.1

# Core dependencies
pyyaml>=6.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
anthropic>=0.39.0

# Already available in environment
streamlit>=1.30.0
python-docx>=1.1.0
jinja2>=3.1.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
